requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
apturl==0.5.2
argcomplete==3.0.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachelib==0.9.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
comm==0.1.3
command-not-found==0.3
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.6
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docopt==0.6.2
docutils==0.19
duplicity==0.8.21
email-validator==1.3.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.0
fasteners==0.14.1
fastjsonschema==2.16.3
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kafka-python==2.0.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mysql-connector-python==8.0.32
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.2.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.1
pytzdata==2020.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.1
reportlab==3.6.8
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.4
rsa==4.9
SecretStorage==3.3.1
Send2Trash==1.8.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.0
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.26.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
unicodecsv==0.14.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.21.1
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.8.2
zipp==1.0.0
